Document compile and run instructions
Documents my experience with compiling and running TAP.
I have documented products that need to be installed, compiling
and running TAP of the command line, as well as using Intellij IDEA.
I have also documented errors and solutions to problems that I
experienced.
The documentation is detailed, but it is meant to save time for
people that come across the same problems in the future.
The documentation is a starting point, and it should
be improved upon as more people download and test the product on
different operating systems.
Another way to improve the documentation would be to do more
automation of some of the build steps if possible.