Upload Service end point
Provide an upload service end point.

Files are loaded into the /tmp directory
A file id is returned to the client which can be later used to locate the loaded file.

The returned id has the format :

randomNumber-extCode-fileSize

The randomNumber is part of the loaded file Name and can be used to locate it.

The extCode is mapped to supported extension codes. Supported extension codes are
txt, jar, zip. All other file types are stored with no extension.

The fileSize does not form part of the uploaded fileName but can be used to verify
the uploaded file.

Here is an example of a request to upload a hello.txt file containing "hello world" message
as issued by GraphQlControllerSpec :

UploadService
[info] application - path=/tmp/TAP-$upload$-$temp$|7576951845638742679|.txt
[info] application - count = 11,status=Success(Done)
[info] application - key=name,filename=hello.txt,contentType=Some(text/plain),file= /tmp/TAP-$upload$-$temp$|7576951845638742679|.txt
[debug] application - size = 11
[debug] application - extCode = 01
[debug] application - randomNumber = 7576951845638742679
[info] application - file id =Some(7576951845638742679-01-11)