Built broken as taking call is missing a parameter
It looks like an earlier change has modified the method this testacase
calls to add a new parameter but the testcase was not modified, or the
testcase modification was not checked in.

Not sure if this modification is correct as the rhetorical moves test
still fails, but it at least gets the build working.